+++ src/pages/SignUp/index.tsx
@@ -1,28 +1,11 @@
-import Button from '@/components/Button';
-import InputButtonSet from '@/components/InputButtonSet';
-import InputText from '@/components/InputText';
-import CommonLayout from '@/components/layout/CommonLayout';
-import ToggleButton from '@/components/ToggleButton';
-import supabase from '@/lib/supabase-client';
-import validator from '@/lib/validator';
-import { useState, useMemo, useCallback, ChangeEvent, useEffect } from 'react';
+import { useState, useCallback, useMemo, useEffect } from 'react';
 import { useNavigate } from 'react-router';
-
-type FormField = 'email' | 'password' | 'confirmPassword' | 'nickname';
-
-interface FormData {
-  email: string;
-  password: string;
-  confirmPassword: string;
-  nickname: string;
-}
-
-interface Errors {
-  emailError: string;
-  passwordError: string;
-  confirmPasswordError: string;
-  nicknameError: string;
-}
+import Button from '@/components/Button';
+import CommonLayout from '@/components/layout/CommonLayout';
+import supabase from '@/lib/supabase-client';
+import { useFormValidation } from './useFormValidation';
+import FormField from './components/FormField';
+import InterestsSection from './components/InterestsSection';
 
 interface Interest {
   id: number;
@@ -34,22 +17,18 @@ const MAX_INTERESTS = 4;
 const SignUp = () => {
   const navigate = useNavigate();
 
-  const [formData, setFormData] = useState<FormData>({
-    email: '',
-    password: '',
-    confirmPassword: '',
-    nickname: '',
-  });
+  const {
+    formData,
+    errors,
+    isEmailValid,
+    isPasswordValid,
+    isConfirmPasswordValid,
+    isNicknameValid,
+    handleChange,
+    checkEmail,
+    checkNickname,
+  } = useFormValidation();
 
-  const [errors, setErrors] = useState<Errors>({
-    emailError: '',
-    passwordError: '',
-    confirmPasswordError: '',
-    nicknameError: '',
-  });
-
-  const [isEmailChecked, setIsEmailChecked] = useState(false);
-  const [isNicknameChecked, setIsNicknameChecked] = useState(false);
   const [interests, setInterests] = useState<Interest[]>([]);
   const [selectedInterests, setSelectedInterests] = useState<string[]>([]);
   const [isSubmitting, setIsSubmitting] = useState(false);
@@ -72,29 +51,6 @@ const SignUp = () => {
     fetchInterests();
   }, []);
 
-  const isEmailValid = useMemo(
-    () => formData.email && (!errors.emailError || errors.emailError === '사용 가능한 이메일입니다.') && isEmailChecked,
-    [formData.email, errors.emailError, isEmailChecked]
-  );
-
-  const isPasswordValid = useMemo(
-    () => formData.password && !errors.passwordError,
-    [formData.password, errors.passwordError]
-  );
-
-  const isConfirmPasswordValid = useMemo(
-    () => formData.confirmPassword && !errors.confirmPasswordError,
-    [formData.confirmPassword, errors.confirmPasswordError]
-  );
-
-  const isNicknameValid = useMemo(
-    () =>
-      formData.nickname &&
-      (!errors.nicknameError || errors.nicknameError === '사용 가능한 닉네임입니다.') &&
-      isNicknameChecked,
-    [formData.nickname, errors.nicknameError, isNicknameChecked]
-  );
-
   const isInterestsValid = useMemo(
     () => selectedInterests.length >= 1 && selectedInterests.length <= MAX_INTERESTS,
     [selectedInterests]
@@ -104,103 +60,6 @@ const SignUp = () => {
     () => isEmailValid && isPasswordValid && isConfirmPasswordValid && isNicknameValid && isInterestsValid,
     [isEmailValid, isPasswordValid, isConfirmPasswordValid, isNicknameValid, isInterestsValid]
   );
-
-  const validateField = useCallback(
-    (name: FormField, value: string) => {
-      let error = '';
-
-      switch (name) {
-        case 'email':
-          if (!validator.isEmail(value)) {
-            error = '올바른 이메일 형식이 아닙니다.';
-          }
-          break;
-        case 'password':
-          if (!validator.isPassword(value)) {
-            error = '영어, 숫자, 특수문자 포함 8자리 이상 입력해주세요.';
-          }
-          if (formData.confirmPassword && value !== formData.confirmPassword) {
-            setErrors((prev) => ({ ...prev, confirmPasswordError: '비밀번호가 일치하지 않습니다.' }));
-          } else {
-            setErrors((prev) => ({ ...prev, confirmPasswordError: '' }));
-          }
-          break;
-        case 'confirmPassword':
-          if (value !== formData.password) {
-            error = '비밀번호가 일치하지 않습니다.';
-          }
-          break;
-        case 'nickname':
-          if (!validator.isNickname(value)) {
-            error = '특수문자 제외 2~8자로 입력해주세요.';
-          }
-          break;
-      }
-
-      setErrors((prev) => ({ ...prev, [`${name}Error`]: error }));
-    },
-    [formData.password, formData.confirmPassword]
-  );
-
-  const handleChange = useCallback(
-    (e: ChangeEvent<HTMLInputElement>) => {
-      const { name, value } = e.target;
-      setFormData((prev) => ({ ...prev, [name]: value }));
-
-      if (name === 'email') {
-        setIsEmailChecked(false);
-      } else if (name === 'nickname') {
-        setIsNicknameChecked(false);
-      }
-
-      validateField(name as FormField, value);
-    },
-    [validateField]
-  );
-
-  const checkEmail = useCallback(async () => {
-    if (!validator.isEmail(formData.email)) return;
-
-    try {
-      const { data, error } = await supabase.from('users').select('email').eq('email', formData.email);
-
-      if (error) throw error;
-
-      if (data && data.length > 0) {
-        setErrors((prev) => ({ ...prev, emailError: '중복된 이메일입니다.' }));
-        setIsEmailChecked(false);
-      } else {
-        setErrors((prev) => ({ ...prev, emailError: '사용 가능한 이메일입니다.' }));
-        setIsEmailChecked(true);
-      }
-    } catch (error) {
-      console.error('Email check error:', error);
-      setErrors((prev) => ({ ...prev, emailError: '이메일 확인 중 오류가 발생했습니다.' }));
-      setIsEmailChecked(false);
-    }
-  }, [formData.email]);
-
-  const checkNickname = useCallback(async () => {
-    if (!validator.isNickname(formData.nickname)) return;
-
-    try {
-      const { data, error } = await supabase.from('users').select('nickname').eq('nickname', formData.nickname);
-
-      if (error) throw error;
-
-      if (data && data.length > 0) {
-        setErrors((prev) => ({ ...prev, nicknameError: '중복된 닉네임입니다.' }));
-        setIsNicknameChecked(false);
-      } else {
-        setErrors((prev) => ({ ...prev, nicknameError: '사용 가능한 닉네임입니다.' }));
-        setIsNicknameChecked(true);
-      }
-    } catch (error) {
-      console.error('Nickname check error:', error);
-      setErrors((prev) => ({ ...prev, nicknameError: '닉네임 확인 중 오류가 발생했습니다.' }));
-      setIsNicknameChecked(false);
-    }
-  }, [formData.nickname]);
 
   const handleToggle = useCallback((interest: string) => {
     setSelectedInterests((prev) => {
@@ -264,35 +123,6 @@ const SignUp = () => {
     }
   }, [formData, isFormValid, navigate, selectedInterests, interests, isSubmitting]);
 
-  const renderError = useCallback((error: string, isSuccess = false) => {
-    return error ? (
-      <p className={`absolute top-full left-0 mt-1.5 text-xs ${isSuccess ? 'text-positive' : 'text-warning'}`}>
-        {error}
-      </p>
-    ) : null;
-  }, []);
-
-  const renderInterestsGrid = useCallback(
-    (startIdx: number, endIdx: number) => {
-      const interestsSlice = interests.slice(startIdx, endIdx);
-
-      return (
-        <div className="grid grid-cols-4 gap-2">
-          {interestsSlice.map((interest) => (
-            <ToggleButton
-              key={interest.id}
-              label={interest.name}
-              isActive={selectedInterests.includes(interest.name)}
-              onClick={() => handleToggle(interest.name)}
-              disabled={selectedInterests.length >= MAX_INTERESTS && !selectedInterests.includes(interest.name)}
-            />
-          ))}
-        </div>
-      );
-    },
-    [interests, selectedInterests, handleToggle]
-  );
-
   return (
     <CommonLayout
       headerProps={{
@@ -300,85 +130,67 @@ const SignUp = () => {
         isLeftIcon: true,
       }}
       showFooter={false}>
+      <div className="flex flex-row justify-between pb-5">
+        <p className="text-primary flex items-center text-base">말랑이에게 정보를 알려주세요</p>
+        <img src="/images/mallang.webp" alt="말랑이" width={50} height={50} />
+      </div>
       <form className="flex flex-col gap-10">
-        <fieldset className="relative">
-          <legend className="sr-only">이메일 입력</legend>
-          <InputButtonSet
-            inputType="text"
-            labelText="이메일"
-            placeholder="이메일 입력"
-            name="email"
-            value={formData.email}
-            onChange={handleChange}
-            onClick={checkEmail}>
-            중복 확인
-          </InputButtonSet>
-          {renderError(errors.emailError, errors.emailError === '사용 가능한 이메일입니다.')}
-        </fieldset>
+        <FormField
+          fieldType="withButton"
+          labelText="이메일"
+          name="email"
+          placeholder="이메일 입력"
+          value={formData.email}
+          error={errors.emailError}
+          isSuccess={errors.emailError === '사용 가능한 이메일입니다.'}
+          onChange={handleChange}
+          onButtonClick={checkEmail}
+          buttonText="중복 확인"
+        />
 
-        <fieldset className="relative">
-          <legend className="sr-only">비밀번호 입력</legend>
-          <InputText
-            labelText="비밀번호"
-            type="password"
-            placeholder="비밀번호 입력"
-            name="password"
-            value={formData.password}
-            onChange={handleChange}
-          />
-          {renderError(errors.passwordError)}
-        </fieldset>
+        <FormField
+          fieldType="password"
+          labelText="비밀번호"
+          name="password"
+          placeholder="비밀번호 입력"
+          value={formData.password}
+          error={errors.passwordError}
+          onChange={handleChange}
+        />
 
-        <fieldset className="relative">
-          <legend className="sr-only">비밀번호 확인</legend>
-          <InputText
-            labelText="비밀번호 확인"
-            type="password"
-            placeholder="비밀번호 확인"
-            name="confirmPassword"
-            value={formData.confirmPassword}
-            onChange={handleChange}
-          />
-          {renderError(errors.confirmPasswordError)}
-        </fieldset>
+        <FormField
+          fieldType="password"
+          labelText="비밀번호 확인"
+          name="confirmPassword"
+          placeholder="비밀번호 확인"
+          value={formData.confirmPassword}
+          error={errors.confirmPasswordError}
+          onChange={handleChange}
+        />
 
-        <fieldset className="relative">
-          <legend className="sr-only">닉네임 입력</legend>
-          <InputButtonSet
-            inputType="text"
-            labelText="닉네임"
-            placeholder="닉네임 입력"
-            name="nickname"
-            value={formData.nickname}
-            onChange={handleChange}
-            onClick={checkNickname}>
-            중복 확인
-          </InputButtonSet>
-          {renderError(errors.nicknameError, errors.nicknameError === '사용 가능한 닉네임입니다.')}
-        </fieldset>
+        <FormField
+          fieldType="withButton"
+          labelText="닉네임"
+          name="nickname"
+          placeholder="닉네임 입력"
+          value={formData.nickname}
+          error={errors.nicknameError}
+          isSuccess={errors.nicknameError === '사용 가능한 닉네임입니다.'}
+          onChange={handleChange}
+          onButtonClick={checkNickname}
+          buttonText="중복 확인"
+        />
 
-        <div className="flex flex-col gap-2">
-          <p className="text-primary text-xs">관심사 (최대 {MAX_INTERESTS}개)</p>
-          {isLoading ? (
-            <p className="text-sm text-gray-500">관심사 목록을 불러오는 중...</p>
-          ) : (
-            <>
-              {interests.length > 0 ? (
-                <>
-                  {renderInterestsGrid(0, 4)}
-                  {interests.length > 4 && renderInterestsGrid(4, 8)}
-                  {interests.length > 8 && renderInterestsGrid(8, 12)}
-                  {interests.length > 12 && renderInterestsGrid(12, 16)}
-                </>
-              ) : (
-                <p className="text-sm text-gray-500">관심사 목록을 불러올 수 없습니다.</p>
-              )}
-            </>
-          )}
-        </div>
+        <InterestsSection
+          interests={interests}
+          selectedInterests={selectedInterests}
+          maxInterests={MAX_INTERESTS}
+          isLoading={isLoading}
+          onToggle={handleToggle}
+        />
       </form>
 
-      <div className="fixed right-0 bottom-4 left-0 mx-auto max-w-[440px] px-4">
+      <div className="mt-42">
         <Button onClick={handleSubmit} ariaDisabled={!isFormValid || isSubmitting || isLoading}>
           말랑이 만나러 가기
         </Button>
